+++ src/components/home/types.ts
@@ -1,0 +1,7 @@
+export interface ISectionElement {
+  title: string;
+  img: string;
+  description: string;
+  style?: object;
+  side?: 'left' | 'right';
+}
